+++ src/state/entity/ig_ActorEntity-base.ts
@@ -3,7 +3,7 @@ import { StateMemory } from '../state-util'
 import * as igAnimatedEntity from './ig_AnimatedEntity-base'
 
 type Return = ReturnType<typeof getEntityState>
-export function getEntityState(this: sc.ActorEntity, _player: StateKey | undefined, memory: StateMemory) {
+export function getEntityState(this: ig.ActorEntity, _player: StateKey | undefined, memory: StateMemory) {
     return {
         ...igAnimatedEntity.getEntityState.call(this, memory),
 
@@ -13,7 +13,7 @@ export function getEntityState(this: sc.ActorEntity, _player: StateKey | undefin
     }
 }
 
-export function setEntityState(this: sc.ActorEntity, state: Return) {
+export function setEntityState(this: ig.ActorEntity, state: Return) {
     igAnimatedEntity.setEntityState.call(this, state)
 
     if (state.animationFixed !== undefined) this.animationFixed = state.animationFixed